requirements.txt
httpx
pydantic
python-dotenv
pyyaml
sqlalchemy
snowflake-connector-python
pandas
pytest
tenacity
schedule
streamlit